+++ packages/pluggableWidgets/maps-web/src/model/tokens.ts
@@ -1,13 +1,17 @@
-import { DerivedPropsGate } from "@mendix/widget-plugin-mobx-kit/main";
 import { token } from "brandi";
+import { DerivedPropsGate } from "@mendix/widget-plugin-mobx-kit/main";
+import { MapsConfig } from "./configs/Maps.config";
+import { CurrentLocationService } from "./services/CurrentLocation.service";
+import { LocationResolverService } from "./services/LocationResolver.service";
+import { MapsSetupService } from "./services/MapsSetup.service";
 import { MapsContainerProps } from "../../typings/MapsProps";
 import { Marker, ModeledMarker } from "../../typings/shared";
-import { MapsConfig } from "./configs/Maps.config";
-import { MapsSetupService } from "./services/MapsSetup.service";
-import { LocationResolverService } from "./services/LocationResolver.service";
 
 /** Function type for geocoding markers. */
 export type GeocodeFunction = (locations?: ModeledMarker[], mapToken?: string) => Promise<Marker[]>;
+
+/** Function type for resolving the current user location. */
+export type GetLocationFunction = () => Promise<Marker>;
 
 /** Tokens to resolve dependencies from the container. */
 
@@ -18,10 +22,12 @@ export const CORE_TOKENS = {
     mainGate: token<DerivedPropsGate<MapsContainerProps>>(label("mainGate")),
     config: token<MapsConfig>(label("config")),
     setupService: token<MapsSetupService>(label("setupService")),
-    geocodeFunction: token<GeocodeFunction>(label("geocodeFunction"))
+    geocodeFunction: token<GeocodeFunction>(label("geocodeFunction")),
+    getLocationFunction: token<GetLocationFunction>(label("getLocationFunction"))
 };
 
 /** Maps-specific tokens. */
 export const MAPS_TOKENS = {
-    locationResolver: token<LocationResolverService>(label("locationResolver"))
+    locationResolver: token<LocationResolverService>(label("locationResolver")),
+    currentLocation: token<CurrentLocationService>(label("currentLocation"))
 };