+++ packages/pluggableWidgets/maps-web/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [5.0.0] - 2026-09-08
+
 ### Fixed
 
 - We fixed an issue where the browser console showed a deprecation warning for every default marker on a Google Maps map.